Move getColorFromItemStack to a different class and other small fixes.
